requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
slowapi>=0.1.9,<1.0
torch>=2.2,<3.0
torchvision>=0.17,<1.0
pillow>=10.0,<12.0
opencv-python-headless>=4.9,<5.0
requests>=2.31,<3.0
numpy>=1.26,<3.0
python-multipart>=0.0.9,<1.0
pytesseract>=0.3.10,<1.0
transformers>=4.40,<5.0
pytest>=8.0,<9.0
httpx>=0.27,<1.0
